+++ log.h
@@ -17,6 +17,7 @@ extern "C" {
 #endif
 
 #include <stdio.h>
+#include <stdint.h>
 
 typedef struct Logger Logger;
 
@@ -32,14 +33,14 @@ Logger *logCreate(void);
  * <port> on host <host>. If the host could not be found, -1 is returned and the
  * channel is not created.
  */
-int logToUDP(Logger *logger, const char *host, int port);
+int logToUDP(Logger *logger, const char *host, uint16_t port);
 
 /*
  * Add an output channel to <logger> that sends messages over a TCP connection
  * to port <port> on host <host>. If no connection could be opened, -1 is
  * returned and the channel is not created.
  */
-int logToTCP(Logger *logger, const char *host, int port);
+int logToTCP(Logger *logger, const char *host, uint16_t port);
 
 /*
  * Add an output channel to <logger> that writes to the file specified with
