+++ net.h
@@ -27,19 +27,19 @@ const char *netHost(uint32_t big_endian_ip);
 /*
  * Get the port that corresponds to service <service>.
  */
-int netPort(const char *service, const char *protocol);
+uint16_t netPort(const char *service, const char *protocol);
 
 /*
  * Bind a socket to <port> and <host>. If <host> is NULL, the socket
  * will be bound to INADDR_ANY. If port is 0, it will be bound to a random port.
  */
-int netBind(int socket, const char *host, int port);
+int netBind(int socket, const char *host, uint16_t port);
 
 /*
  * Connect an existing socket <fd> to <host> and <port>. Returns 0 on success or -1 if an error
  * occurs.
  */
-int netConnect(int fd, const char *host, int port);
+int netConnect(int fd, const char *host, uint16_t port);
 
 /*
  * Get hostname of peer.
@@ -49,7 +49,7 @@ const char *netPeerHost(int sd);
 /*
  * Get port number used by peer.
  */
-int netPeerPort(int sd);
+uint16_t netPeerPort(int sd);
 
 /*
  * Get local hostname.
@@ -59,7 +59,7 @@ const char *netLocalHost(int sd);
 /*
  * Get local port number.
  */
-int netLocalPort(int sd);
+uint16_t netLocalPort(int sd);
 
 #ifdef __cplusplus
 }
