+++ ns.c
@@ -4,7 +4,7 @@
  * Part of libjvs.
  *
  * Copyright:	(c) 2013 Jacco van Schaik (jacco@jaccovanschaik.net)
- * Version:	$Id: ns.c 277 2016-10-07 08:51:20Z jacco $
+ * Version:	$Id: ns.c 291 2018-05-07 11:58:12Z jacco $
  *
  * This software is distributed under the terms of the MIT license. See
  * http://www.opensource.org/licenses/mit-license.php for details.
@@ -12,6 +12,7 @@
 
 #include <unistd.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 #include <errno.h>
 
@@ -33,42 +34,42 @@ static void ns_handle_data(Dispatcher *dis, int fd, void *udata)
 
     int n;
 
-P   dbgPrint(stderr, "Reading from fd %d...\n", fd);
+    P dbgPrint(stderr, "Reading from fd %d...\n", fd);
 
     n = read(fd, data, sizeof(data));
 
-P   dbgPrint(stderr, "read() returned %d.\n", n);
+    P dbgPrint(stderr, "read() returned %d.\n", n);
 
     if (n > 0) {
-P       dbgPrint(stderr, "Adding to incoming buffer.\n");
+        P dbgPrint(stderr, "Adding to incoming buffer.\n");
 
         bufAdd(&conn->incoming, data, n);
 
         if (ns->on_socket_cb != NULL) {
-P           dbgPrint(stderr, "Calling on_socket_cb.\n");
+            P dbgPrint(stderr, "Calling on_socket_cb.\n");
 
             ns->on_socket_cb(ns, fd,
                     bufGet(&conn->incoming), bufLen(&conn->incoming), ns->on_socket_udata);
         }
     }
     else if (n == 0) {
-P       dbgPrint(stderr, "End of file, disconnecting.\n");
+        P dbgPrint(stderr, "End of file, disconnecting.\n");
 
         nsDisconnect(ns, fd);
 
         if (ns->on_disconnect_cb != NULL) {
-P           dbgPrint(stderr, "Calling on_disconnect_cb.\n");
+            P dbgPrint(stderr, "Calling on_disconnect_cb.\n");
 
             ns->on_disconnect_cb(ns, fd, ns->on_disconnect_udata);
         }
     }
     else {
-P       dbgPrint(stderr, "Error, disconnecting.\n");
+        P dbgPrint(stderr, "Error, disconnecting.\n");
 
         nsDisconnect(ns, fd);
 
         if (ns->on_error_cb != NULL) {
-P           dbgPrint(stderr, "Calling on_error_cb.\n");
+            P dbgPrint(stderr, "Calling on_error_cb.\n");
 
             ns->on_error_cb(ns, fd, errno, ns->on_error_udata);
         }
@@ -81,7 +82,7 @@ static void ns_add_connection(NS *ns, int fd)
 
     paSet(&ns->connections, fd, conn);
 
-P   dbgPrint(stderr, "New connection on fd %d\n", fd);
+    P dbgPrint(stderr, "New connection on fd %d\n", fd);
 
     disOnData(&ns->dis, fd, ns_handle_data, NULL);
 }
@@ -126,7 +127,7 @@ NS *nsCreate(void)
  * will be accepted automatically. Data coming in on the resulting socket will be reported via the
  * callback installed using nsOnSocket().
  */
-int nsListen(NS *ns, const char *host, int port)
+int nsListen(NS *ns, const char *host, uint16_t port)
 {
     int listen_fd;
 
@@ -179,11 +180,11 @@ void nsOnError(NS *ns, void (*cb)(NS *ns, int fd, int error, void *udata), void 
 }
 
 /*
- * Make a connection to the given <host> and <port>. Incoming data on this socket is reported using
- * the callback installed with nsOnSocket(). The new file descriptor is returned, or -1 if an error
- * occurred.
+ * Make a connection to the given <host> and <port>. Incoming data on this 
+ * socket is reported using the callback installed with nsOnSocket(). The new 
+ * file descriptor is returned, or -1 if an error occurred.
  */
-int nsConnect(NS *ns, const char *host, int port)
+int nsConnect(NS *ns, const char *host, uint16_t port)
 {
     int fd = tcpConnect(host, port);
 
@@ -201,7 +202,7 @@ void nsDisconnect(NS *ns, int fd)
 {
     NS_Connection *conn = paGet(&ns->connections, fd);
 
-P   dbgPrint(stderr, "Closing fd %d\n", fd);
+    P dbgPrint(stderr, "Closing fd %d\n", fd);
 
     close(fd);
 
@@ -229,7 +230,7 @@ void nsOnData(NS *ns, int fd, void (*cb)(NS *ns, int fd, void *udata), const voi
  */
 void nsDropData(NS *ns, int fd)
 {
-P   dbgPrint(stderr, "Calling disDropData for fd %d\n", fd);
+    P dbgPrint(stderr, "Calling disDropData for fd %d\n", fd);
 
     disDropData(&ns->dis, fd);
 }
@@ -448,11 +449,11 @@ int errors = 0;
 
 void on_time(NS *ns, double t, void *udata)
 {
-    int port = *((int *) udata);
+    uint16_t port = *((uint16_t *) udata);
     static int fd, step = 0;
 
-P   dbgPrint(stderr, "port = %d\n", port);
-P   dbgPrint(stderr, "step = %d\n", step);
+    P dbgPrint(stderr, "port = %d\n", port);
+    P dbgPrint(stderr, "step = %d\n", step);
 
     switch(step) {
     case 0:
@@ -471,7 +472,7 @@ P   dbgPrint(stderr, "step = %d\n", step);
     nsOnTime(ns, t + 0.1, on_time, udata);
 }
 
-void tester(int port)
+void tester(uint16_t port)
 {
     int r;
 
@@ -481,19 +482,19 @@ void tester(int port)
 
     r = nsRun(ns);
 
-P   dbgPrint(stderr, "nsRun returned %d\n", r);
+    P dbgPrint(stderr, "nsRun returned %d\n", r);
 
     nsDestroy(ns);
 }
 
 void on_connect(NS *ns, int fd, void *udata)
 {
-P   dbgPrint(stderr, "fd = %d\n", fd);
+    P dbgPrint(stderr, "fd = %d\n", fd);
 }
 
 void on_disconnect(NS *ns, int fd, void *udata)
 {
-P   dbgPrint(stderr, "fd = %d\n", fd);
+    P dbgPrint(stderr, "fd = %d\n", fd);
 
     nsClose(ns);
 }
@@ -512,7 +513,7 @@ void on_socket(NS *ns, int fd, const char *data, int size, void *udata)
 
     nsDiscard(ns, fd, n);
 
-P   dbgPrint(stderr, "fd = %d, str = \"%s\", n = %d\n", fd, str, n);
+    P dbgPrint(stderr, "fd = %d, str = \"%s\", n = %d\n", fd, str, n);
 }
 
 void testee(NS *ns)
@@ -525,7 +526,7 @@ void testee(NS *ns)
 
     r = nsRun(ns);
 
-P   dbgPrint(stderr, "nsRun returned %d\n", r);
+    P dbgPrint(stderr, "nsRun returned %d\n", r);
 
     nsDestroy(ns);
 }
@@ -535,10 +536,10 @@ int main(int argc, char *argv[])
     NS *ns = nsCreate();
 
     int listen_fd   = nsListen(ns, "localhost", 0);
-    int listen_port = netLocalPort(listen_fd);
+    uint16_t listen_port = netLocalPort(listen_fd);
 
-P   fprintf(stderr, "listen_fd = %d\n", listen_fd);
-P   fprintf(stderr, "listen_port = %d\n", listen_port);
+    P fprintf(stderr, "listen_fd = %d\n", listen_fd);
+    P fprintf(stderr, "listen_port = %d\n", listen_port);
 
     if (fork() == 0) {
         /* Child. */
